Split `PurchasesTests` step 10: extracted `PurchasesGetProductsTests`
